style: add background to copy svg refactor: prefix blog posts url with `/blog`
